Move `structure.xyz` loading to a separate function
- Previously, the code was inside SCBA, but now is inside a function in `inputs.py`.
- This is done to be able to nicely use the function in tests.
